SR/requirements.txt
pandas==1.5.2
numpy==1.23.5
scikit-surprise==1.1.1
gradio
